HyperShift capability with S3 chart creation.
Forked from github.com/validatedpatterns/multicloud-gitops and removed:
- ACM
- config-demo application
- hello-world application

HyperShift requires that the multicluster-engine operator is installed
w/the hypershift TP feature enabled. Additionally, an S3 bucket must
be created for OIDC.

values-secret-template.yaml has been updated with a consumable secret
that will be used with vault and eso to create the secrets required
for both ack-s3 and hypershift projects.

Lastly, within the hypershift application an imagestream is created in
the multicluster-engine namespace that can be used to provision clusters.
Create a job using the imagestream to take advantage of the hypershift cli.